+++ webcomponents/src/SimpleTooltip/index.ts
@@ -1,0 +1,1 @@
+export { default as SimpleTooltip, SimpleTooltipProps } from './SimpleTooltip';