+++ webcomponents/src/index.ts
@@ -5,6 +5,7 @@ export * from './Logo';
 export * from './Modal';
 export * from './ModalContent';
 export * from './SearchBar';
+export * from './SimpleTooltip';
 export * from './Table';
 export * from './TableBodyCell';
 export * from './TableHeaderCell';